fix: remove duplicate infrastructure/logger.ts and move logger to lib/
Delete the leftover src/infrastructure/ directory containing a duplicate
logger.ts, and relocate logger from core/ to lib/ since it has side
effects (console.*) which is inconsistent with core/'s pure types and
functions.
