+++ src/index.ts
@@ -7,7 +7,7 @@ import { resolve } from "node:path";
 import { cancel, intro, log, outro, spinner } from "@clack/prompts";
 import { cli, define } from "gunshi";
 import pc from "picocolors";
-import { logger } from "./core/logger.js";
+import { logger } from "./lib/logger.js";
 import { initMarkdown } from "./lib/markdown.js";
 import { isNodeError } from "./lib/node-error.js";
 import { resolveStyles } from "./lib/styles.js";
