--- src/infrastructure/logger.ts
@@ -1,7 +1,0 @@
-const PREFIX = "[peek]";
-
-export const logger = {
-  info: (...args: unknown[]) => console.log(PREFIX, ...args),
-  warn: (...args: unknown[]) => console.warn(PREFIX, ...args),
-  error: (...args: unknown[]) => console.error(PREFIX, ...args),
-} as const;
