+++ src/server/routes/directory.tsx
@@ -5,9 +5,9 @@ import { MarkdownContent } from "../../components/layout/markdown-content.js";
 import { PageHeader } from "../../components/layout/page-header.js";
 import { Sidebar } from "../../components/navigation/sidebar.js";
 import type { FileTreeNode } from "../../core/file-tree.js";
-import { logger } from "../../core/logger.js";
 import { isWithinBase } from "../../core/path.js";
 import type { FileTreeCache } from "../../lib/file-tree-cache.js";
+import { logger } from "../../lib/logger.js";
 import { renderMarkdown } from "../../lib/markdown.js";
 import { readTextFile } from "../../lib/read-text-file.js";
 import type { ResolvedStyles } from "../../lib/styles.js";
